Data Science/Data Science Projects/Assignment_3_WizardBooks_API/requirements.txt
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
numpy==1.26.4
pymongo==4.13.0
pyodbc==5.1.0
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.1.3
